feat: package latest released tempest and plugins
This patch aims to revive the snap-tempest project. Starting from this
change, the snapcraft tool will include the latest released tempest and
supported plugins from pypi upon a triggered build. Since currently the
latest released OpenStack version is 2023.1/Antelope, the plugin list is
taken from the Antelope release page[0].

According to the Tempest and Plugins compatible version policy[1], this
release should be compatible with all the releases with Maintained
status. As of today, these releases are:
- Bobcat
- Antelope
- Zed
- Yoga

Therefore the snap built today can be released to the <release>/edge
channels for these releases alongside with latest/edge.

Once a release reachs EM phase, newly built snaps will not be released
to its channel anymore.

[0] https://releases.openstack.org/antelope/#tempest-plugins
[1] https://docs.openstack.org/tempest/latest/tempest_and_plugins_compatible_version_policy.html

Signed-off-by: Mert Kırpıcı <mert.kirpici@canonical.com>