feat: introduce automation for stable branches
This commit introduces a github workflow for automating updating the
requirements for each of the openstack releases, based on their release
page metadata taken from the opendev.org/openstack/releases repository.

To introduce a new release:
- a stable/<release> branch needs to be created
- the update-tempest-releases github workflow needs to be modified

In addition to that, starting from this commit, PR workflow will run
against the changes to the stable/* branches as well as master and main.