Allow manually bumping the Tempest version in snapcraft.yaml (#78)
update_snapcraft.py is used to automatically detect when there are
available updates for component packaged in the snap (Tempest,
tempestconf, Tempest plugins). Since the comparison is a simple diff, it
is impossible to manually override versions in snapcraft.yaml because
the automation would try to revert the changes. This commit prevents
reverting to an older Tempest version, therefore allowing for manual
version bumps (of Tempest only - the behavior remains unchanged for
other components)

Partial: #76