Move manually injected requirements to a separate file - Ussuri.
This commit is based off commit
a803b203180d7167d7f6665a4b9f1935aab7b3c7, and adapted for the
Ussuri-specific depenency on confluent-kafka==1.2.0

This commit moves a single requirement that used to be part of
update_snapcraft.py in a separate file. This makes update_snapcraft.py
completely independent of the target OpenStack release.