Determine the latest compatible version via RSS feeds.
The tempest and tempest plugin metadata files in the releases[0] repository are
not updated continuously and therefore do not reflect the actual latest
compatible version for a given OpenStack release (see also the Tempest
nad Plugins compatible version policy[1]).
This commit switches to using the individual projects' RSS feeds as
source of truth as to which version should be picked; specifically, the
"{release}-last" tag if available, and the latest version tag otherwise.

[0] https://opendev.org/openstack/releases.git
[1] https://docs.openstack.org/tempest/latest/tempest_and_plugins_compatible_version_policy.html

Closes: #79