Move the release tox environment to the main branch
This commit splits tox.ini between the more common environments (which
will continue existing in all branches), and the release one, which is
only going to be used to update tempest in snapcraft.yaml. The release
environment can then be kept in the main branch only, along with the
related update script.

Closes: #85