Remove watcher from excluded-plugins
commit 550e43bb45 in python-temepstconf introduces
discovery of watcher service and updates tempest
config accordingly. The changes from v3.5.0 to the
above mentioned commit are few that should not break
anything on caracal. The changes introdcued are case
sensitive for tempest config options and discovery
of watcher service. So install python-tempestconf to
the above commit in the snap to include watcher
discoverability feature.

Add excluded-plugins.txt from snap-tempest-automation
project and remove watcher-tempest-plugin from the
list.